requirements-gpu.txt
tensorflow-gpu==2.8.0
opencv-python==4.5.5.64
lxml==4.8.0
tqdm==4.64.0
absl-py==1.0.0
easydict==1.9
matplotlib==3.5.2
pillow==9.1.0
pandas==1.4.2
